+++ EdgarClerc/CW1/Services/StudentService.cs
@@ -89,7 +89,7 @@ public class StudentService
         int totalStudents,
         int totalGrades,
         double meanOfMeans,
-        int maxGrade,
+        float maxGrade,
         bool hasFailing,
         bool allHaveEmail
     ) GetStatistics()
